Split urban_heat_island.rs into urban_heat_island/ subdirectory
Split crates/simulation/src/urban_heat_island.rs (657 lines) into focused modules
under urban_heat_island/ for better maintainability:
- constants.rs (42 lines): UHI constants and surface heat factors
- types.rs (65 lines): UhiGrid resource and effective_temperature helper
- calculations.rs (86 lines): is_nighttime, surface_heat_factor, local_green_fraction
- systems.rs (149 lines): update_uhi_grid system and UrbanHeatIslandPlugin
- tests.rs (336 lines): all unit tests
- mod.rs (12 lines): thin facade with re-exports
