+++ crates/simulation/src/urban_heat_island/mod.rs
@@ -1,0 +1,12 @@
+pub mod calculations;
+pub mod constants;
+pub mod systems;
+pub mod types;
+
+#[cfg(test)]
+mod tests;
+
+pub use systems::update_uhi_grid;
+pub use systems::UrbanHeatIslandPlugin;
+pub use types::effective_temperature;
+pub use types::UhiGrid;
